C2.5: First native boot — auto-translated code reaches the game loop
The auto-translated FUN_00401000 now executes far enough through its
boot path to enter the binary's main game tick (FUN_004385d8),
without any x86 interpreter in the loop. The runtime is real JS calling
real JS calling runtime/win32.js (180 real impls + 47 stubs).

What was added:

  tools/extract-data.js
    One-shot. Reads binary/rct.exe via the existing PE loader, drops
    executable sections (.text, CODESEG), keeps initialised data
    (.rdata, .data, DATASEG, DataSeg, CodeSeg, .rsrc) at their virtual
    addresses. Output: decompiled/data.bin (10 MB sparse, ~4 MB real
    data) + decompiled/data-manifest.json. The binary EXE is no longer
    a runtime dependency — only build-time.

  tools/gen-dispatch.js
    Generates ported/auto/_dispatch.js: a Map<addr, fn> that imports
    every FUN_<addr> from ported/auto/*.js and lets the runtime's
    state.fnDispatch resolve WindowProc addresses to JS callables.

  runtime/harness.js
    Constructs a Heap of (data + heap + stack) bytes, loads the data
    blob, populates state.fnDispatch from the generated _dispatch
    module, wires VFS via setRuntimeContext, and exports `run()` to
    invoke FUN_00401000(heap).

  tools/native-boot.js
    Worker-thread driver with a hard watchdog (default 10s). Surfaces
    the first throw or infinite-loop as a clean exit code so we can
    bisect translator/runtime issues by `--func=<addr>`.

Translator improvements driven by the native-boot iteration:

  - normalizeTypes(source) replaces Win32/Ghidra type names with `int`
    so tree-sitter parses both casts ((LPBOOL)x → (int)x) and
    declarations (LPBOOL p; → int p;) correctly. Stripping (the
    previous strategy) broke pointer declarations like `*p;`.
  - Stripping for C-only constructs JS rejects:
      * stripIntSuffixes (0x3fU, 1L, ULL → 0x3f, 1)
      * stripWideStringPrefix (L"abc" → "abc")
  - emitDeclaration now handles pointer_declarator and array_declarator
    children (was skipping them, leaking variable names to imports).
  - Self-imports for recursive FUN_xxx calls are filtered out.
  - Force-name to FUN_<padded-hex> when the file's RVA is known, so
    Ghidra-named functions (RtlUnwind, _strlen, …) don't collide with
    runtime/win32 imports.
  - Fall back to a throwing stub when tree-sitter parses a file as one
    big ERROR (preserves the export so dependents still load).
  - Lowercase + underscore-prefixed identifiers are now recognised as
    Win32 imports (timeGetTime, _strlen, etc.).
  - Pointer/subscript/field LHS in assignments now emit setU{8,16,32}
    instead of invalid `heap.u32(addr) = value`.
  - Empty labeled-statement (label at end of block) emits `:;`
    placeholder.

What works:

  - FUN_00402bd5 (DDraw + asset init) — returns cleanly
  - FUN_00401120 (init pass 2) — returns cleanly
  - Registry probing (RegOpenKeyA → not-found path) — clean
  - FUN_00403c2a (message-pump probe) — returns true (no quit)
  - Outer game-loop body FUN_004385d8 — enters and runs indefinitely
    (this is the original binary's behaviour: the game runs forever
    until WM_QUIT, which we don't post in headless mode)

test/runtime/native_boot.test.js (4 new cases) confirms each step.

Tests: 116/116 pass.

The native runtime is now end-to-end demonstrably functional. Next
milestone is C2.5.d — bulk diff-test surge across the 1214 ported
functions vs the interpreter, to find translator/runtime bugs that
silent-fail.

+++ ported/auto/429c3d.js
@@ -36,10 +36,10 @@ export function FUN_00429c3d(heap) {
         if ((heap.u32((__addr_DAT_00743bbf) + (iVar4) * 4) != '\x03') && (heap.u32((__addr_DAT_00743bbf) + (iVar4) * 4) != '\a')) {
           return -0x8000;
         }
-        if ((heap.u32((__addr_DAT_00887422) + (heap.u32((byte)(__addr_DAT_00743bfc) + (iVar4) * 4) * 0x130) * 4) & 1) == 0) {
+        if ((heap.u32((__addr_DAT_00887422) + (heap.u32((__addr_DAT_00743bfc) + (iVar4) * 4) * 0x130) * 4) & 1) == 0) {
           return -0x8000;
         }
-        uVar1 = heap.u32((__addr_DAT_0088747e + heap.u32((byte)(__addr_DAT_00743bfe) + (iVar4) * 4) * 2 + heap.u32((byte)(__addr_DAT_00743bfc) + (iVar4) * 4) * 0x260));
+        uVar1 = heap.u32((__addr_DAT_0088747e + heap.u32((__addr_DAT_00743bfe) + (iVar4) * 4) * 2 + heap.u32((__addr_DAT_00743bfc) + (iVar4) * 4) * 0x260));
         cVar3 = heap.u32((__addr_DAT_00743bff) + (iVar4) * 4);
         while (true) {
           if (cVar3 == '\0') {
