+++ ported/auto/440143.js
@@ -10,7 +10,7 @@ export function FUN_00440143(heap) {
   const __addr_DAT_00743bbf = __sp + 4;
   try {
   let in_EAX = 0;
-  (heap.u32(heap.u32((__addr_PTR_LAB_0062d44c) + (heap.u32((byte)(__addr_DAT_00743bbf) + ((in_EAX & 0xffff) * 0x100) * 4)) * 4)))();
+  (heap.u32(heap.u32((__addr_PTR_LAB_0062d44c) + (heap.u32((__addr_DAT_00743bbf) + ((in_EAX & 0xffff) * 0x100) * 4)) * 4)))();
   return;
 } finally {
     heap.freeFrame(8);
